Extracted the HighlightImage to a new component
It's better, and more reusable this way. I'm going to use it for Bloopy,
and maybe the other projects too.
