+++ src/Components/Highlight/stylish.css
@@ -86,9 +86,6 @@
     position: relative;
     overflow: hidden;
 
-    /* z-index: 1; */
-    /* transform: translateZ(0); */
-
     /* border-radius: 5px; */
 }
 
@@ -96,7 +93,6 @@
     position: absolute;
 
     bottom: 0;
-    /* border-radius: 0; */
 
     height: 100%;
     width: 100%;
@@ -113,7 +109,6 @@ h3.highlight-title {
     padding: 0;
 
     bottom: 5%;
-    /* top: 10%; */
     left: 5%;
 
     border-bottom: none;
